+++ example/lib/main.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+import 'package:iconnect/iconnect.dart';
+import 'package:example/app/app.dart';
+
+void main() {
+  runApp(provider(MyApp()));
+}